pyproject.toml
[project]
name = "orchestral"
version = "1.0.0"
description = "AI Model Orchestration Platform - Unified interface for ChatGPT, Claude, and Gemini"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "Orchestral Team", email = "team@orchestral.ai"}
]
keywords = ["ai", "llm", "chatgpt", "claude", "gemini", "orchestration", "multi-model"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "openai>=1.40.0",
    "anthropic>=0.34.0",
    "google-generativeai>=0.8.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "httpx>=0.27.0",
    "tenacity>=9.0.0",
    "structlog>=24.4.0",
    "python-dotenv>=1.0.0",
    "tiktoken>=0.7.0",
    "rich>=13.8.0",
    "typer>=0.12.0",
    "aiofiles>=24.1.0",
    "prometheus-client>=0.21.0",
    "python-multipart>=0.0.9",
    "sse-starlette>=2.1.0",
    "PyYAML>=6.0.0",
    "pillow>=10.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "pytest-mock>=3.14.0",
    "mypy>=1.11.0",
    "ruff>=0.6.0",
    "pre-commit>=3.8.0",
    "httpx>=0.27.0",
]

[project.scripts]
orchestral = "orchestral.cli:app"

[project.urls]
Homepage = "https://github.com/orchestral/orchestral"
Documentation = "https://docs.orchestral.ai"
Repository = "https://github.com/orchestral/orchestral"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/orchestral"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --cov=src/orchestral --cov-report=term-missing"

[tool.ruff]
target-version = "py311"
line-length = 100
select = ["E", "F", "W", "I", "N", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.ruff.isort]
known-first-party = ["orchestral"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
